fix type of `len` to size_t
git-svn-id: https://svn.r-project.org/R/trunk@81039 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/modules/internet/sockconn.c
@@ -200,7 +200,7 @@ static size_t sock_read(void *ptr, size_t size, size_t nitems, Rconnection con)
 static size_t sock_write(const void *ptr, size_t size, size_t nitems, Rconnection con)
 {
     Rsockconn this = (Rsockconn)con->private;
-    ssize_t n = R_SockWrite(this->fd, ptr, (int)(size * nitems), this->timeout) / ((ssize_t)size);
+    ssize_t n = R_SockWrite(this->fd, ptr, (size_t)(size * nitems), this->timeout) / ((ssize_t)size);
     return n > 0 ? n : 0;
 }
